% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_metadata}
\alias{add_metadata}
\title{Add additional global metadata.}
\usage{
add_metadata(rsmith, ...)
}
\arguments{
  \item{rsmith}{An \code{\link{rsmith}} object.}

  \item{...}{Additional named arguments to add to
  metadata/}
}
\value{
An modified \code{\link{rsmith}} object.
}
\description{
Add additional global metadata.
}
\examples{
static_site <- rsmith_demo("static-site")
add_metadata(static_site, test = 1, foo = "bar")
}
\seealso{
Other plugins: \code{\link{ignore_drafts}};
  \code{\link{ignore_files}}
}

