% Generated by roxygen2 (4.0.0): do not edit by hand
\name{brew}
\alias{brew}
\title{Render brew templates.}
\usage{
brew(pattern = "\\\\.html$", template_dir = "templates")
}
\arguments{
  \item{pattern}{Regular expression describing patterns to
  process.}

  \item{template_dir}{Directory in which to look for
  templates. Relative to site base directory.}
}
\description{
Any files matching \code{pattern}, and containing a template entry
in their metadata will have their contents replaced by the template
processed by \code{\link[brew]{brew}}.
}
\details{
Metadata in the page metadata is available as is. Site metadata
is available with the prefix \code{site.}.
}
\examples{
brew_site <- rsmith_demo("brew")
brew_site \%>\% use(brew()) \%>\% preview()
}

