% Generated by roxygen2 (4.0.0): do not edit by hand
\name{build}
\alias{build}
\alias{preview}
\alias{watch}
\title{Save rsmith output.}
\usage{
build(rsmith)

preview(rsmith)

watch(rsmith, interval = 0.25)
}
\arguments{
  \item{rsmith}{An object created by \code{\link{rsmith}}.}

  \item{interval}{How often to check for changes, in
  seconds.}
}
\description{
\code{build} will build the site once, while \code{watch} listens for
changes, re-outputing every time an input file changes. \code{preview}
prints output on screen.
}
\examples{
static_site <- rsmith_demo("static-site")
static_site \%>\% build()
\dontrun{
static_site \%>\% watch()
}
}

