% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collections}
\alias{collections}
\title{Create collections of files}
\usage{
collections(pattern, sortby = "date", reverse = TRUE)
}
\arguments{
\item{pattern}{Regular expression to select files into the collection.}

\item{sortby}{Field to sort by.}

\item{reverse}{Whether to sort in decreasing order.}
}
\description{
Finds all files that match a given pattern, sorts them
according to some criteria and adds 'next' and 'previous' links
to them.
}
\examples{
rmarkdown_demo <- rsmith_demo("rmarkdown")
rmarkdown_demo \%>\%
  use(collections(pattern=".Rmd")) \%>\%
  use(rmarkdown()) \%>\%
  preview()
}

