% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ignore_files}
\alias{ignore_files}
\title{Ignore files that match a pattern.}
\usage{
ignore_files(rsmith, pattern, ...)
}
\arguments{
  \item{rsmith}{An \code{\link{rsmith}} object.}

  \item{pattern}{A regular expression to match against the
  file paths.  Any files matching the pattern will be
  dropped.}

  \item{...}{Additional arguments passed to
  \code{\link{grepl}}}
}
\value{
An modified \code{\link{rsmith}} object.
}
\description{
Ignore files that match a pattern.
}
\examples{
static_site <- rsmith_demo("static-site")
ignore_files(static_site, "first")
ignore_files(static_site, "post")
}
\seealso{
Other plugins: \code{\link{add_metadata}};
  \code{\link{remove_drafts}}
}

