% Generated by roxygen2 (4.0.1): do not edit by hand
\name{markdown}
\alias{markdown}
\title{Convert markdown to html}
\usage{
markdown(pattern = "\\\\.md$")
}
\arguments{
\item{pattern}{Regular expression used to identify markdown files.}
}
\description{
Use the \pkg{markdown} package. Markdown inputs are converted to html
and their extension changed to \code{.html}. Binary files are ignored.
}
\examples{
static_site <- rsmith_demo("static-site")
static_site \%>\% preview()
static_site \%>\% use(markdown()) \%>\% preview()
}

