% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plugin}
\alias{plugin}
\alias{plugin_with_init}
\title{Create your own plugin.}
\usage{
plugin(name, process)

plugin_with_init(name, init, process)
}
\arguments{
  \item{name}{of the plugin}

  \item{process}{function called to process each file}

  \item{init}{function called once with site metadata}
}
\description{
Create your own plugin.
}

