% Generated by roxygen2 (4.0.0): do not edit by hand
\name{remove_drafts}
\alias{remove_drafts}
\title{Remove draft file.}
\usage{
remove_drafts(rsmith)
}
\arguments{
  \item{rsmith}{An \code{\link{rsmith}} object.}
}
\value{
An modified \code{\link{rsmith}} object.
}
\description{
Any file with \code{draft: true} in the metadata will be removed.
}
\examples{
static_site <- rsmith_demo("static-site")
static_site$files[[1]]$metadata$draft <- TRUE
remove_drafts(static_site)
}
\seealso{
Other plugins: \code{\link{add_metadata}};
  \code{\link{ignore_files}}
}

