% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rmarkdown}
\alias{rmarkdown}
\title{Convert rmarkdown to html}
\usage{
rmarkdown(pattern = "\\\\.Rmd$")
}
\arguments{
  \item{pattern}{Regular expression used to identify
  Rmarkdown files.}
}
\description{
Use the \pkg{rmarkdown} package. You can set options that apply to all
files in the global metadata field \code{rmarkdown}.
}
\examples{
rmarkdown_demo <- rsmith_demo("rmarkdown")
rmarkdown_demo \%>\% preview()
rmarkdown_demo \%>\% use(rmarkdown()) \%>\% preview()
}

