% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rsmith}
\alias{rsmith}
\title{Initialise rsmith.}
\usage{
rsmith(src = "src", dest = "build", base = ".")
}
\arguments{
  \item{src,dest}{Source and destination directories}

  \item{base}{Base directory - should usually be set to the
  parent of the src.}
}
\description{
RSmith is a static site generator for R. This function provides
basic metadata about the site. You then \code{\link{add_plugin}}s to do
manipulate the source files, and finally use \code{\link{build}} or
\code{\link{watch}} to save the results to disk.
}
\examples{
rsmith_demo("static-site") \%>\% build()
}

