% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rsmith_demo}
\alias{rsmith_demo}
\title{Easily access rsmith demo sites.}
\usage{
rsmith_demo(name, ...)
}
\arguments{
  \item{name}{Name of demo.}
}
\description{
Easily access rsmith demo sites.
}
\examples{
dir(system.file("examples", package = "rsmith"))
rsmith_demo("static-site")
}
\keyword{internal}

