% Generated by roxygen2 (4.0.0): do not edit by hand
\name{transform_each}
\alias{transform_each}
\title{\code{transform_each()} is a helper to make it easier to write functions
that operate on files independently. More complex plugins should operate
directly on the rsmith object.}
\usage{
transform_each(rsmith, plugin)
}
\arguments{
  \item{plugin}{A plugin is a function with two arguments,
  which returns a modified file or NULL.}
}
\description{
\code{transform_each()} is a helper to make it easier to write functions
that operate on files independently. More complex plugins should operate
directly on the rsmith object.
}

