% Generated by roxygen2 (4.0.0): do not edit by hand
\name{use}
\alias{use}
\title{Add a plugin.}
\usage{
use(rsmith, plugin)
}
\arguments{
  \item{rsmith}{\code{\link{rsmith}} object}

  \item{plugin}{object created by \code{\link{plugin}}}
}
\value{
a modified rsmith \code{\link{rsmith}} object.
}
\description{
You can use one of the pre-built functions or create your own with
\code{\link{plugin}}.
}
\seealso{
Other plugins: \code{\link{add_metadata}};
  \code{\link{ignore_drafts}}; \code{\link{ignore_files}}
}

