% Generated by roxygen2 (4.0.0): do not edit by hand
\name{use_markdown}
\alias{use_markdown}
\title{Convert markdown to html}
\usage{
use_markdown(pattern = "\\\\.md$")
}
\arguments{
  \item{pattern}{Regular expression used to identify
  markdown files.}
}
\description{
Use the \pkg{markdown} package. Markdown inputs are converted to html
and their extension changed to \code{.html}.
}
\examples{
static_site <- rsmith_demo("static-site")
static_site \%>\% preview()
static_site \%>\% add_plugin(use_markdown()) \%>\% preview()
}

