% Generated by roxygen2 (4.0.0): do not edit by hand
\name{use_whisker}
\alias{use_whisker}
\title{Render whisker templates.}
\usage{
use_whisker(rsmith, pattern = "\\\\.R?md$", template_dir = "templates")
}
\description{
Any files matching \code{pattern}, and containing a template entry
in their metadata will have their contents replaced by the template
processed by \code{\link[whisker]{whisker.render}}.
}
\details{
Metadata in the page metadata is available as is. Site metadata
is available with the prefix \code{site.}.

Uses moustache syntax: \url{http://mustache.github.io/}.
}

