% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ignore_drafts}
\alias{ignore_drafts}
\title{Remove draft file.}
\usage{
ignore_drafts()
}
\value{
An \code{\link{plugin}}
}
\description{
Any file with \code{draft: true} in the metadata will be removed.
}
\examples{
static_site <- rsmith_demo("static-site")
static_site \%>\% add_plugin(ignore_drafts()) \%>\% preview()
}
\seealso{
Other plugins: \code{\link{add_metadata}};
  \code{\link{add_plugin}}; \code{\link{ignore_files}}
}

