% Generated by roxygen2 (4.0.0): do not edit by hand
\name{use_whisker}
\alias{use_whisker}
\title{Render whisker templates.}
\usage{
use_whisker(rsmith, pattern = "\\\\.R?md$", template_dir = "templates")
}
\arguments{
  \item{rsmith}{xxx}

  \item{pattern}{Regular expression describing patterns to
  process.}

  \item{template_dir}{Directory in which to look for
  templates. Relative to site base directory.}
}
\description{
Any files matching \code{pattern}, and containing a template entry
in their metadata will have their contents replaced by the template
processed by \code{\link[whisker]{whisker.render}}.
}
\details{
Metadata in the page metadata is available as is. Site metadata
is available with the prefix \code{site.}.

Uses moustache syntax: \url{http://mustache.github.io/}.
}
\examples{
static_site <- rsmith_demo("static-site")
static_site \%>\% preview()
static_site \%>\% add_plugin(use_whisker()) \%>\% preview()
}

